+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - Fixed SettingsValidator to accept both deprecated and new settings data structure
 - Enabled notifications on periodic movements
+- Improved performance of On-Edit Trigger
 
 ## [0.4.6](https://github.com/bueckerlars/obsidian-note-mover-shortcut/compare/0.4.5...0.4.6)